+++ generator/compile_typst.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python3
+
+import subprocess
+import sys
+from pathlib import Path
+
+def main():
+    pass
