+++ parser/parser.py
@@ -6,7 +6,7 @@ from typing import Any
 
 from pydantic import BaseModel, Field
 
-from generator import GeneralLLM
+from generator.generator import GeneralLLM
 
 
 class ParsedEvent(BaseModel):
